+++ plugins/bundler.ts
@@ -1,20 +1,31 @@
 import { merge } from "../utils.ts";
+import Site from "../site.ts";
+import { Page } from "../filesystem.ts";
+
+interface Options {
+  extensions: string[];
+  sourceMap: boolean;
+  options: Deno.EmitOptions;
+}
 
 // Default options
-const defaults = {
+const defaults: Options = {
   extensions: [".ts", ".js"],
   sourceMap: false,
   options: {},
 };
 
-export default function (userOptions = {}) {
+/**
+ * Plugin to load all .js and .ts files and bundle them using Deno.emit
+ */
+export default function (userOptions: Partial<Options>) {
   const options = merge(defaults, userOptions);
 
-  return (site) => {
+  return (site: Site) => {
     site.loadAssets(options.extensions);
     site.process(options.extensions, bundler);
 
-    async function bundler(file) {
+    async function bundler(file: Page) {
       const from = site.src(file.src.path + file.src.ext);
       const { files } = await Deno.emit(from, {
         ...options,