+++ plugins/date.ts
@@ -1,5 +1,7 @@
 import { format } from "../deps/date.ts";
 import { merge } from "../utils.ts";
+import Site from "../site.ts";
+import { Helper } from "../types.ts";
 
 const formats = new Map([
   ["ATOM", "yyyy-MM-dd'T'HH:mm:ssXXX"],
@@ -10,20 +12,32 @@ const formats = new Map([
   ["HUMAN_DATETIME", "PPPppp"],
 ]);
 
+interface Options {
+  locales: Record<string, unknown>;
+  formats: Record<string, string>;
+}
+
 // Default options
-const defaults = {
+const defaults: Options = {
   locales: {},
   formats: {},
 };
 
-export default function (userOptions = {}) {
+/**
+ * Plugin to format Date values
+ */
+export default function (userOptions: Partial<Options>) {
   const options = merge(defaults, userOptions);
   const defaultLocale = Object.keys(options.locales).shift();
 
-  return (site) => {
-    site.filter("date", filter);
+  return (site: Site) => {
+    site.filter("date", filter as Helper);
 
-    function filter(date, pattern = "DATE", lang = defaultLocale) {
+    function filter(
+      date: string | Date,
+      pattern = "DATE",
+      lang = defaultLocale,
+    ) {
       if (!date) {
         return;
       }
@@ -34,7 +48,7 @@ export default function (userOptions = {}) {
 
       const patt = options.formats[pattern] || formats.get(pattern) ||
         pattern;
-      const locale = options.locales[lang];
+      const locale = lang ? options.locales[lang] : undefined;
 
       return format(date, patt, { locale });
     }