+++ plugins/eta.ts
@@ -15,6 +15,9 @@ const defaults: Options = {
   includes: "",
 };
 
+/**
+ * Plugin to add support for Eta as a template engine
+ */
 export default function (userOptions: Partial<Options>) {
   return (site: Site) => {
     const options = merge(