+++ plugins/svgo.ts
@@ -1,25 +1,39 @@
+import Site from "../site.ts";
 import { optimize } from "../deps/svgo.ts";
 import { merge } from "../utils.ts";
+import { Page } from "../filesystem.ts";
+
+interface Options {
+  extensions: string[];
+  options: {
+    [index: string]: unknown;
+  };
+}
 
 // Default options
-const defaults = {
+const defaults: Options = {
   extensions: [".svg"],
   options: {},
 };
 
-export default function (userOptions) {
+/**
+ * Plugin to load all .svg files and minify them
+ * using SVGO
+ */
+export default function (userOptions: Partial<Options>) {
   const options = merge(defaults, userOptions);
 
-  return (site) => {
+  return (site: Site) => {
     site.loadAssets(options.extensions);
     site.process(options.extensions, svg);
 
-    async function svg(page) {
+    async function svg(page: Page) {
       const path = site.src(page.dest.path + page.dest.ext);
       const result = await optimize(page.content, {
         path,
         ...options.options,
-      });
+      }) as { data: string };
+
       page.content = result.data;
     }
   };