+++ plugins/terser.ts
@@ -32,6 +32,10 @@ const defaults: Options = {
   },
 };
 
+/**
+ * Plugin to load all .js files and minify them
+ * using Terser
+ */
 export default function (userOptions: Partial<Options>) {
   const options = merge(defaults, userOptions);